batch: Move realized entry storage
Realized batch entries expose a line view model together with compact
provenance storage in one module. Merge, discard, and boundary helpers
share that module when they need storage-aware accessors.

The project does not provide a named boundary for the compact storage
contract. Batch helpers depend on underscored names from the entry-view
module, so callers need to know which parts are storage mechanics and which
parts describe one realized line.

This commit addresses that by moving compact storage, streaming, and
storage-aware accessor helpers into realized_entry_storage, leaving
realized_entries to define RealizedEntry. Source callers now import
RealizedEntries and storage helpers from the storage module.

Subsequent commits will update behavior tests to import the new storage
boundary and add architecture coverage for the split.
