+++ src/git_stage_batch/batch/realized_boundaries.py
@@ -4,12 +4,12 @@ from __future__ import annotations
 
 from collections.abc import Sequence
 
-from .realized_entries import (
-    RealizedEntry as _RealizedEntry,
-    _RealizedEntries,
-    _entry_content_at,
-    _entry_is_claimed_at,
-    _entry_source_line_at,
+from .realized_entries import RealizedEntry as _RealizedEntry
+from .realized_entry_storage import (
+    RealizedEntries,
+    realized_entry_content_at,
+    realized_entry_is_claimed_at,
+    realized_entry_source_line_at,
 )
 from ..exceptions import (
     AmbiguousAnchorError as _AmbiguousAnchorError,
@@ -32,7 +32,7 @@ def find_realization_fallback_boundary(
         return 0
 
     prior_source_line: int | None = None
-    if isinstance(entries, _RealizedEntries):
+    if isinstance(entries, RealizedEntries):
         for run in entries.provenance_runs():
             if run.source_start == 0:
                 continue
@@ -45,7 +45,7 @@ def find_realization_fallback_boundary(
                 prior_source_line = max(prior_source_line or candidate, candidate)
     else:
         for index in range(len(entries)):
-            entry_source_line = _entry_source_line_at(entries, index)
+            entry_source_line = realized_entry_source_line_at(entries, index)
             if entry_source_line is not None and entry_source_line < source_line:
                 prior_source_line = max(
                     prior_source_line or entry_source_line,
@@ -69,7 +69,7 @@ def find_boundary_after_source_line(
     matching_indices = []
     claimed_indices = []
 
-    if isinstance(entries, _RealizedEntries):
+    if isinstance(entries, RealizedEntries):
         for run in entries.provenance_runs():
             if run.source_start == 0:
                 continue
@@ -82,9 +82,9 @@ def find_boundary_after_source_line(
                 claimed_indices.append(index)
     else:
         for i in range(len(entries)):
-            if _entry_source_line_at(entries, i) == source_line:
+            if realized_entry_source_line_at(entries, i) == source_line:
                 matching_indices.append(i)
-                if _entry_is_claimed_at(entries, i):
+                if realized_entry_is_claimed_at(entries, i):
                     claimed_indices.append(i)
 
     if not matching_indices:
@@ -124,7 +124,7 @@ def boundary_choices_after_source_line(
         return (0,)
 
     matching_indices: list[int] = []
-    if isinstance(entries, _RealizedEntries):
+    if isinstance(entries, RealizedEntries):
         for run in entries.provenance_runs():
             if run.source_start == 0:
                 continue
@@ -134,7 +134,7 @@ def boundary_choices_after_source_line(
             matching_indices.append(run.dest_start + (source_line - run.source_start))
     else:
         for index in range(len(entries)):
-            if _entry_source_line_at(entries, index) == source_line:
+            if realized_entry_source_line_at(entries, index) == source_line:
                 matching_indices.append(index)
 
     if not matching_indices:
@@ -158,7 +158,7 @@ def sequence_present_at_boundary(
         return False
 
     return all(
-        _normalize_line_content(_entry_content_at(entries, boundary + i))
+        _normalize_line_content(realized_entry_content_at(entries, boundary + i))
         == _normalize_line_endings(sequence[i])
         for i in range(len(sequence))
     )
